requirements.txt
python 3.9.9
pyaudio 
openai_whisper 
torch
numpy
scipy